% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bLOO.R
\name{bLOO_checks}
\alias{bLOO_checks}
\title{Input data checks for bLOO}
\usage{
bLOO_checks(tpoints, radius, min_train)
}
\arguments{
\item{tpoints}{sf point object. Contains the training points samples.}

\item{radius}{Numeric. Positive radius of the exclusion buffer (in the same units as the tpoints).}

\item{min_train}{Numeric, between 0 and 1. Proportion of minimum training
data that must be used in each CV fold. Defaults to 0 (i.e. no restrictions). If
positive and not met for a given iteration, the radius will be decreased by
1/10 until the condition is met.}
}
\description{
Input data checks for bLOO
}
