% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndm.R
\name{nndm_checks}
\alias{nndm_checks}
\title{Input data checks for NNDM}
\usage{
nndm_checks(tpoints, ppoints, phi, min_train)
}
\arguments{
\item{tpoints}{sf or sfc point object. Contains the training points samples.}

\item{ppoints}{sf or sfc point object. Contains the prediction target points.}

\item{phi}{Numeric. Estimate of the landscape autocorrelation range (in the
same units as the tpoints and ppoints).}

\item{min_train}{Numeric between 0 and 1. Minimum proportion of training
data that must be used in each CV fold. Defaults to 0 (i.e. no restrictions).}
}
\description{
Input data checks for NNDM
}
