% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bLOO.R
\name{plot.bLOO}
\alias{plot.bLOO}
\title{Plot method for \emph{bLOO} class}
\usage{
\method{plot}{bLOO}(x, it, ...)
}
\arguments{
\item{x}{An object of type \emph{bLOO}.}

\item{it}{bLOO iteration that needs to be plotted.}

\item{...}{other arguments.}
}
\description{
Plot method for \emph{bLOO} class
}
