% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndm.R
\name{plot.nndm}
\alias{plot.nndm}
\title{Plot method for \emph{nndm} class}
\usage{
\method{plot}{nndm}(x, ...)
}
\arguments{
\item{x}{An object of type \emph{nndm}.}

\item{...}{other arguments.}
}
\description{
Plot method for \emph{nndm} class
}
