% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bLOO.R
\name{print.bLOO}
\alias{print.bLOO}
\title{Print method for \emph{bLOO} class}
\usage{
\method{print}{bLOO}(x, ...)
}
\arguments{
\item{x}{An object of type \emph{bLOO}.}

\item{...}{other arguments.}
}
\description{
Print method for \emph{bLOO} class
}
