% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndm.R
\name{print.nndm}
\alias{print.nndm}
\title{Print method for \emph{nndm} class}
\usage{
\method{print}{nndm}(x, ...)
}
\arguments{
\item{x}{An object of type \emph{nndm}.}

\item{...}{other arguments.}
}
\description{
Print method for \emph{nndm} class
}
